+++ lib/src/methods/sdk.dart
@@ -2,7 +2,6 @@ import 'package:likeminds_feed/src/methods/helper.dart';
 import 'package:likeminds_feed/src/methods/methods.dart';
 import 'package:likeminds_feed/src/repositories/access_repository.dart';
 import 'package:likeminds_feed/src/repositories/auth_repository.dart';
-import 'package:likeminds_feed/src/repositories/branding_repository.dart';
 import 'package:likeminds_feed/src/repositories/helper_repository.dart';
 import 'package:likeminds_feed/src/repositories/media_repository.dart';
 import 'package:likeminds_feed/src/repositories/post_repository.dart';
@@ -50,14 +49,6 @@ class SdkApplication {
     return MediaApi(
       mediaRepository: GetIt.instance.get<MediaRepository>(
         instanceName: DIService.kInstanceMediaRepository,
-      ),
-    );
-  }
-
-  BrandingApi getBrandingApi() {
-    return BrandingApi(
-      brandingRepository: GetIt.instance.get<BrandingRepository>(
-        instanceName: DIService.kInstanceBrandingRepository,
       ),
     );
   }